+++ core/dbt/include/global_project/macros/etc/get_custom_database.sql
@@ -14,6 +14,10 @@
 
 #}
 {% macro generate_database_name(custom_database_name=none, node=none) -%}
+    {% do return(adapter_macro('generate_database_name', custom_database_name, node)) %}
+{%- endmacro %}
+
+{% macro default__generate_database_name(custom_database_name=none, node=none) -%}
     {%- set default_database = target.database -%}
     {%- if custom_database_name is none -%}